connect wallet step replaced
