+++ src/pages/Ownership.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+
+export function Ownership() {
+  return (
+    <div className="flex h-full flex-col items-center pb-5">
+      <h1 className="text-2xl font-bold mb-4">Ownership Management</h1>
+      <p>Ownership management page content will go here</p>
+    </div>
+  );
+}
